requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
torch
torchvision
ultralytics
opencv-python
requests
boto3
python-dotenv
python-multipart